requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI client for Whisper STT, TTS, and GPT
openai==1.3.7

# Google Generative AI (Gemini) for evaluation
google-generativeai==0.3.2

# Core Python utilities
requests==2.31.0
python-dotenv==1.0.0
typing-extensions==4.8.0

# MongoDB support
pymongo==4.6.0
motor==3.3.2

# Audio processing support
Pillow==10.1.0
numpy==1.24.4

# Optional: For enhanced audio format support
soundfile==0.12.1
librosa==0.10.1

# Voice metrics analysis
scipy==1.11.4
python-speech-features==0.6

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2